+++ src/ui.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 
-export let juleVersionStatus: vscode.StatusBarItem | undefined = undefined;
+export let juleVersionStatus: vscode.StatusBarItem | null = null;
 
 export function setJuleVersionStatus(item: vscode.StatusBarItem): void {
 	juleVersionStatus = item;